+++ generators/datetime/index.js
@@ -1,4 +1,6 @@
 const {iterateCLDRData, generateImport, generateDataModule, convertJSONValueToAST} = require('../utils');
+const logger = require('../logger');
+
 
 /**
  * @param {string} locale
@@ -98,11 +100,9 @@ const generator = {
 	 * @return {string}
 	 */
 	generateFormats(locales) {
+		logger.verbose(`Generating time formats for: ${locales.join(', ')}`);
 		const data = {};
 
-		// TODO: This iterates locales exactly as specified, but should attempt to reduce them same reduceLocale in lib
-		// i.e. If locale is set in config as "ru-RU" this will fail to produce data set both for "ru-RU" and "ru".
-		// Ditto other generators.
 		for (const [locale, localeData] of iterateCLDRData('cldr-dates-modern', 'ca-gregorian.json', locales)) {
 			data[locale] = convertJSONValueToAST(processTimeData(locale, localeData));
 		}
@@ -117,6 +117,7 @@ const generator = {
 	 * @return {string}
 	 */
 	generateRelativeFormats(locales) {
+		logger.verbose(`Generating relative time formats for: ${locales.join(', ')}`);
 		const data = {};
 
 		for (const [locale, localeData] of iterateCLDRData('cldr-dates-modern', 'dateFields.json', locales)) {