+++ generators/logger.js
@@ -1,0 +1,3 @@
+const {logger} = require('zombiebox');
+
+module.exports = logger.createChild('i18n');